requirements.txt
requests==2.22.0
Django==2.2.1
djangorestframework==3.10.3
gunicorn==20.0.4
django-cors-middleware==1.4.0
mysqlclient==1.4.2
openpyxl==3.0.0
Pillow==7.1.2



